requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==24.10.0
boto3==1.36.2
botocore==1.36.2
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.5
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
isort==5.13.2
jmespath==1.0.1
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pycodestyle==2.12.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
